% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov_ker}
\alias{cov_ker}
\title{Kernel estimated covariance when the index variable equals newu
x: observed predictors of the training sample.
u: observed index variable of the training sample.
h_cov: bandwidth for variance estimation.
h_mean: bandwidth for mean estimation}
\usage{
cov_ker(x, u, newu, h_cov, h_mean = NULL)
}
\description{
Kernel estimated covariance when the index variable equals newu
x: observed predictors of the training sample.
u: observed index variable of the training sample.
h_cov: bandwidth for variance estimation.
h_mean: bandwidth for mean estimation
}
