% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_dspca.R
\name{cv_dspca}
\alias{cv_dspca}
\title{Cross-validation for DSPCA}
\usage{
cv_dspca(
  x,
  u,
  y,
  n_folds = 5,
  type = "lda",
  h_mean = (1:15)/10,
  h_cov = (1:15)/10,
  rho = exp(-1:6),
  k = 1:5
)
}
\arguments{
\item{x}{the input matrix. Each column consists of observations
from an input variable.}

\item{u}{a numeric vector, consisting of observations from the
index variable.}

\item{y}{a two-level factor, consisting of observations from the
response variable. Levels represent class labels.}

\item{n_folds}{number of folds used for cross-validation.}

\item{type}{the classification method used in conjunction with the
dynamic supervised principal component analysis framework.
Supported methods are \code{type = "lda"} for linear discriminant
analysis and \code{type = "qda"} for quadratic discriminant analysis.}

\item{h_mean}{a numeric vector of kernel bandwidths for the
Nadaraya-Watson estimator of the means, to be chosen by
leave-one-out cross-validation.}

\item{h_cov}{a numeric vector of kernel bandwidths for the
Nadaraya-Watson estimator of the covariances, to be chosen by
leave-one-out cross-validation.}

\item{rho}{a numeric vector of weight parameters for constructing
the total covariance, to be chosen by cross-validation.}

\item{k}{the number of eigenvectors used for projection.}
}
\value{
An object of S3 class \code{"cv_dspca"}, which is a list with
the following components:

\item{x1}{the submatrix of \code{x} with observations from the first
class.}

\item{x2}{the submatrix of \code{x} with observations from the second
class.}

\item{u1}{the subvector of \code{u} with observations from the first
class.}

\item{u2}{the subvector of \code{u} with observations from the second
class.}

\item{h1_mean}{the optimal kernel bandwidth for the
Nadaraya-Watson estimator of the first mean chosen by
leave-one-out cross-validation.}

\item{h2_mean}{the optimal kernel bandwidth for the
Nadaraya-Watson estimator of the second mean chosen by
leave-one-out cross-validation.}

\item{h1_cov}{the optimal kernel bandwidth for the Nadaraya-Watson
estimator of the first covariance chosen by leave-one-out
cross-validation.}

\item{h2_cov}{the optimal kernel bandwidth for the Nadaraya-Watson
estimator of the second covariance chosen by leave-one-out
cross-validation.}

\item{rho_opt}{the optimal weight parameter for constructing the
total covariance chosen by cross-validation.}

\item{k_opt}{the optimal number of eigenvectors used for
projection chosen by cross-validation.}

\item{lvs}{the two levels of the factor \code{y}.}
}
\description{
Runs K-fold cross-validation for DSPCA.
}
\details{
Runs K-fold cross-validation to select the optimal values of
\code{h_mean}, \code{h_cov}, \code{rho} and \code{k}. Returns an object of class
\code{"cv_dspca"} that contains the optimal tuning parameters,
including the bandwiths for Nadaraya-Watson estimator, the weight
in the total covariance, and the number of eigenvectors used for
projection. Call the \code{predict} function with the returned object
to predict the class label of new observations.
}
\author{
Wenbo Ouyang, Ruiyang Wu, Ning Hao, Hao Helen Zhang
}
