% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_h_cov}
\alias{cv_h_cov}
\title{Leave one out cross validation for choosing the best bandwidth in
the kernel estimation of the covariance.
h_cov: a range of bandwidths to be chosen from
u: observed index variable of the training sample.
x: observed predictors of the training sample.
h_mean: the bandwidth for mean estimation.}
\usage{
cv_h_cov(h_cov, x, u, h_mean)
}
\description{
Leave one out cross validation for choosing the best bandwidth in
the kernel estimation of the covariance.
h_cov: a range of bandwidths to be chosen from
u: observed index variable of the training sample.
x: observed predictors of the training sample.
h_mean: the bandwidth for mean estimation.
}
