% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_h_mean}
\alias{cv_h_mean}
\title{Leave one out cross validation for choosing the best bandwidth in
the kernel estimation of the mean.
h_mean: a range of bandwidths to be chosen from
x: observed predictors of the training sample.
u: observed index variable of the training sample.}
\usage{
cv_h_mean(h_mean, x, u)
}
\description{
Leave one out cross validation for choosing the best bandwidth in
the kernel estimation of the mean.
h_mean: a range of bandwidths to be chosen from
x: observed predictors of the training sample.
u: observed index variable of the training sample.
}
