% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_ker}
\alias{mean_ker}
\title{Kernel estimated mean when the index variable equals newu
x: observed predictors of the training sample.
u: observed index variable of the training sample.
h: bandwidth.}
\usage{
mean_ker(x, u, newu, h)
}
\description{
Kernel estimated mean when the index variable equals newu
x: observed predictors of the training sample.
u: observed index variable of the training sample.
h: bandwidth.
}
