% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_dspca.R
\name{predict.cv_dspca}
\alias{predict.cv_dspca}
\title{DSPCA Prediction}
\usage{
\method{predict}{cv_dspca}(object, newx, newu, type = "lda")
}
\arguments{
\item{object}{fitted object from the \code{cv_dspca} function.}

\item{newx}{a matrix of new values of \code{x}. Each row is a new
observation whose class label is to be precdicted.}

\item{newu}{a vector of new values of the index variable \code{u}.}

\item{type}{the classification method used in conjunction with the
dynamic supervised principal component analysis framework.
Supported methods are \code{type = "lda"} for linear discriminant
analysis and \code{type = "qda"} for quadratic discriminant analysis.}
}
\value{
Predicted class labels for \code{newx} and \code{newu}.
}
\description{
This function is used to predict the class labels for new
observations
}
\details{
\code{predict.cv_dspca} outputs the predicted class labels for new
observations. The new values of the input variables are passed in
\code{newx}. The new values of the index variable are passed in \code{newu}.
}
\author{
Wenbo Ouyang, Ruiyang Wu, Ning Hao, Hao Helen Zhang
}
